back/requirements.txt
Django==2.2.10
psycopg2==2.8.2
djangorestframework==3.9.4
django-filter==2.1.0
elasticsearch-dsl>=6.0,<6.2
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-cors-headers




